+++ CHANGELOG.md
@@ -19,6 +19,7 @@ incremented for features.
 
 * ts: Use `hex` by default for decoding Instruction ([#547](https://github.com/project-serum/anchor/pull/547)).
 * lang: `CpiAccount::reload` mutates the existing struct instead of returning a new one ([#526](https://github.com/project-serum/anchor/pull/526)).
+* cli: `[clusters.<network>]` Anchor.toml section has been renamed to `[programs.<network>]` ([#]()).
 
 ## [0.11.1] - 2021-07-09